App.js 와 AppContext.js 분리
