+++ frontend/src/App.js
@@ -1,10 +1,12 @@
-import { createContext, useState, useEffect } from 'react';
+import { useContext } from 'react';
 import { Route, Routes, useLocation } from 'react-router-dom';
+import { AppContext } from './AppContext';
 import Modal from 'react-modal';
 import MainPage from './pages/Main/MainPage';
 import LogInPage from './pages/user/LogIn/LogInPage';
 import SignUpPage from './pages/user/SignUp/SignUpPage';
 import Header from './components/section/Header/Header';
+import Aside from './components/section/Aside/Aside';
 import WardPage from './pages/Ward/WardPage';
 import ViewConversationPage from './pages/guardians/ViewConversation/ViewConversationPage';
 import ViewEmotionAnalysisPage from './pages/guardians/ViewEmotionAnalysis/ViewEmotionAnalysisPage';
@@ -15,47 +17,8 @@ import WardSettingsPage from './pages/guardians/WardSettings/WardSettingsPage';
 
 Modal.setAppElement('#root');
 
-const AppContext = createContext();
-
-const AppProvider = ({ children }) => {
-  const [isLoggedIn, setIsLoggedIn] = useState(false);
-  const [isEdit, setIsEdit] = useState(false);
-  const [isWardSetting, setIsWardSetting] = useState(false);
-  const [isGuardian, setIsGuardian] = useState(false);
-  const [isHelpOpen, setIsHelpOpen] = useState(false);
-  const [isCalendarOpen, setIsCalendarOpen] = useState(false);
-  const [isSummaryOpen, setIsSummaryOpen] = useState(false);
-  const [helpStep, setHelpStep] = useState(0);
-  const [isEmojiSelected, setIsEmojiSelected] = useState(false);
-
-  useEffect(() => {
-    const storedLoggedInState = sessionStorage.getItem('isLoggedIn');
-    if (storedLoggedInState === 'true') {
-        setIsLoggedIn(true);
-    }
-
-  }, []);
-
-  return (
-    <AppContext.Provider 
-      value={{
-        isLoggedIn, setIsLoggedIn, 
-        isEdit, setIsEdit, 
-        isWardSetting, setIsWardSetting, 
-        isGuardian, setIsGuardian,
-        isHelpOpen, setIsHelpOpen,
-        isCalendarOpen, setIsCalendarOpen,
-        isSummaryOpen, setIsSummaryOpen,
-        helpStep, setHelpStep,
-        isEmojiSelected, setIsEmojiSelected,
-      }}
-    >
-      {children}
-    </AppContext.Provider>
-  );
-}
-
 function App() {
+  const { isGuardian } = useContext(AppContext);
   const location = useLocation();
 
   const getPageAddress = () => {
@@ -79,23 +42,23 @@ function App() {
   }
 
   return (
-    <AppProvider>
-        <Header pageAddress={getPageAddress()}/>
-        <Routes>
-          <Route path='/' element={<MainPage/>}/>
-          <Route path='/LogInPage' element={<LogInPage/>}/>
-          <Route path='/SignUpPage' element={<SignUpPage/>}/>
-          <Route path='/WardPage' element={<WardPage/>}/>
-          <Route path='/ViewConversationPage' element={<ViewConversationPage/>}/>
-          <Route path='/ViewEmotionAnalysisPage' element={<ViewEmotionAnalysisPage/>}/>
-          <Route path='/SchedulePage' element={<SchedulePage/>}/>
-          <Route path='/HomeInformationSettingsPage' element={<HomeInformationSettingsPage/>}/>
-          <Route path='/DodamSettingsPage' element={<DodamSettingsPage/>}/>
-          <Route path='/WardSettingsPage' element={<WardSettingsPage/>}/>
-        </Routes>
-    </AppProvider>
+    <>
+      <Header pageAddress={getPageAddress()}/>
+      {isGuardian && <Aside/>}
+      <Routes>
+        <Route path='/' element={<MainPage/>}/>
+        <Route path='/LogInPage' element={<LogInPage/>}/>
+        <Route path='/SignUpPage' element={<SignUpPage/>}/>
+        <Route path='/WardPage' element={<WardPage/>}/>
+        <Route path='/ViewConversationPage' element={<ViewConversationPage/>}/>
+        <Route path='/ViewEmotionAnalysisPage' element={<ViewEmotionAnalysisPage/>}/>
+        <Route path='/SchedulePage' element={<SchedulePage/>}/>
+        <Route path='/HomeInformationSettingsPage' element={<HomeInformationSettingsPage/>}/>
+        <Route path='/DodamSettingsPage' element={<DodamSettingsPage/>}/>
+        <Route path='/WardSettingsPage' element={<WardSettingsPage/>}/>
+      </Routes>
+    </>
   );
 }
 
 export default App;
-export { AppContext, AppProvider };
