+++ frontend/src/components/guardians/Conversation/ConversationBoard.js
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import { AppContext } from '../../../App';
+import { AppContext } from '../../../AppContext';
 
 const convertToKST = (utcTime) => {
     const date = new Date(`1970-01-01T${utcTime}Z`);
