+++ frontend/src/components/guardians/EmotionAnalysis/EmotionAnalysisBoard.js
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import { AppContext } from '../../../App';
+import { AppContext } from '../../../AppContext';
 
 const EmotionAnalysisBoard = ({ testBoard }) => {
     const { isHelpOpen, helpStep } = useContext(AppContext);
