+++ frontend/src/components/guardians/EmotionAnalysis/EmotionAnalysisGraph.js
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import { AppContext } from '../../../App';
+import { AppContext } from '../../../AppContext';
 
 const EmotionAnalysisGraph = ({ testGraph, realKey }) => {
     const { isHelpOpen, helpStep } = useContext(AppContext);
