+++ frontend/src/components/guardians/Guardian/Help/Help.js
@@ -2,7 +2,7 @@ import React, { useContext, useEffect } from 'react';
 import { FiXCircle, FiArrowLeftCircle, FiArrowRightCircle } from "react-icons/fi";
 import Modal from 'react-modal';
 import HelpContents from './HelpContents';
-import { AppContext } from '../../../../App';
+import { AppContext } from '../../../../AppContext';
 
 const Help = ({ pageAddress }) => {
     const { isHelpOpen, setIsHelpOpen, setIsCalendarOpen,
