+++ frontend/src/components/guardians/Guardian/Toggle/Toggle.js
@@ -1,29 +1,26 @@
 import { useNavigate, useLocation } from 'react-router-dom';
-import React, { useEffect, useState } from 'react';
+import React, { useContext, useEffect, useState } from 'react';
 import './Toggle.css';
+import { AppContext } from '../../../../AppContext';
 
 const Toggle = () => {
+    const { setIsGuardian } = useContext(AppContext);
     const [isToggled, setIsToggled] = useState('피보호자');
 
     const navigate = useNavigate();
     const location = useLocation();
 
     useEffect(() => {
-        if (location.pathname === '/WardSettingsPage') {
-            setIsToggled('보호자');
-        } else if (location.pathname === '/WardPage') {
+        if (location.pathname === '/WardPage') {
             setIsToggled('피보호자');
+            setIsGuardian(false);
         }
     }, [location.pathname]);
 
     const modeChange = (e) => {
         const mode = e.target.value;
         setIsToggled(mode);
-        if (mode === '피보호자') {
-            navigate('/WardPage');
-        } else {
-            navigate('/WardSettingsPage');
-        }
+        navigate(mode==='보호자' ? '/ViewConversationPage' : '/WardPage');
     };
 
     return (
