+++ frontend/src/components/section/Aside/Aside.js
@@ -1,9 +1,9 @@
-import React, { useEffect } from 'react';
+import React from 'react';
 import { Link, useLocation } from 'react-router-dom';
 import '../../Wave.css';
 
 const Aside = () => {
-    const features = [
+    const menu = [
         { 
             title: "대화 내용 보기", 
             url: "/ViewConversationPage",
@@ -38,27 +38,23 @@ const Aside = () => {
 
     const location = useLocation();
 
-    const imageLocation = (features) => {
-        if (location.pathname === features.url) {
+    const imageLocation = (menu) => {
+        if (location.pathname === menu.url) {
             return "./images/orange_star.png";
         } else {
             return "./images/star.png";
         }
     };
 
-    useEffect(()=> {
-        console.log(location.pathname)
-    }, [location.pathname])
-
     return (
-        <aside className='fixed left-0 top-0 h-screen shadow-[1px_0px_6px_#a5996e] z-1 bg-primary'>
+        <aside className='fixed left-0 top-0 h-screen shadow-[1px_0px_6px_#a5996e] z-2 bg-primary overflow-auto'>
             <nav className='flex flex-col items-start justify-between mt-[80px] h-[90vh]'>
-                {features.map((features, index) => (
-                    <Link key={index} to={features.url}> 
+                {menu.map((menu, index) => (
+                    <Link key={index} to={menu.url}> 
                         <div className='wave inline-flex items-center relative m-7 text-middle-size'>
-                            <img className='flex w-[60px] h-[60px]' src={imageLocation(features)}/>
-                            <img className='relative -left-[43px] top-1 w-[25px] h-[25px] mr-[-20px]' src={features.image}/>
-                            <h2 className='wave'>{features.title}</h2>
+                            <img className='flex w-[60px] h-[60px]' src={imageLocation(menu)}/>
+                            <img className='relative -left-[43px] top-1 w-[25px] h-[25px] mr-[-20px]' src={menu.image}/>
+                            <h2 className='wave'>{menu.title}</h2>
                         </div>
                     </Link>
                 ))}
