+++ frontend/src/pages/Ward/WardPage.js
@@ -1,14 +1,7 @@
-import React, { useContext, useEffect } from 'react';
+import React from 'react';
 import SpeechToText from '../../components/Ward/SpeechToText';
-import { AppContext } from '../../App';
 
 const WardPage = () => {
-    const { setIsGuardian } = useContext(AppContext);
-    
-    useEffect(() => {
-        setIsGuardian(false);
-    }, [setIsGuardian])
-
     return (
         <div>
             <SpeechToText/>
