+++ frontend/src/pages/guardians/DodamSettings/DodamSettingsPage.js
@@ -1,9 +1,8 @@
 import React, { useContext, useState, useEffect }  from 'react';
-import Aside from '../../../components/section/Aside/Aside';
 import api from '../../../Service/Api';
 import DodamVoiceForm from '../../../components/guardians/DodamSettings/DodamVoiceForm';
 import Guardian from '../../../components/guardians/Guardian/Guardian';
-import { AppContext } from '../../../App';
+import { AppContext } from '../../../AppContext';
 
 const DodamSettingsPage = () => {
     const { isGuardian, setIsGuardian, isWardSetting } = useContext(AppContext);
@@ -53,7 +52,6 @@ const DodamSettingsPage = () => {
 
     return (
         <div className='flex flex-col h-screen w-screen pl-[240px] overflow-x-hidden'>
-            <Aside/>
             <div className='pt-28 pl-5'>
                 <h2 className='text-3xl text-left'>도담이 설정</h2>
                 <div>
