+++ frontend/src/pages/guardians/HomeInformationSettings/HomeInformationSettingsPage.js
@@ -1,9 +1,8 @@
 import React, { useState, useContext } from 'react';
-import Aside from '../../../components/section/Aside/Aside';
-import { AppContext } from '../../../App';
 import { FiTrash2, FiEdit2 } from "react-icons/fi";
 import HomeInformationForm from '../../../components/guardians/HomeInformationSettings/HomeInformationModal/HomeInformationForm';
 import Guardian from '../../../components/guardians/Guardian/Guardian';
+import { AppContext } from '../../../AppContext';
 
 const HomeInformationSettingsPage = () => {
     const { isGuardian, setIsGuardian, isWardSetting } = useContext(AppContext);
@@ -36,7 +35,6 @@ const HomeInformationSettingsPage = () => {
 
     return (
         <div className='flex flex-col h-screen w-screen pl-[240px] text-2xl'>
-            <Aside/>
             <div className='pt-28 pl-5'>
                 <h2 className='text-3xl text-left'>집 정보 설정</h2>
                 <div className='grid grid-cols-1'>
