+++ frontend/src/pages/guardians/Schedule/SchedulePage.js
@@ -1,9 +1,8 @@
 import React, { useContext, useState } from 'react';
 import { FiEdit2, FiTrash2 } from "react-icons/fi";
-import Aside from '../../../components/section/Aside/Aside';
 import ScheduleForm from '../../../components/guardians/Schedule/ScheduleForm';
 import Guardian from '../../../components/guardians/Guardian/Guardian';
-import { AppContext } from '../../../App';
+import { AppContext } from '../../../AppContext';
 
 const SchedulePage = () => {
     const { isGuardian, setIsGuardian, isWardSetting, isHelpOpen, helpStep } = useContext(AppContext);
@@ -43,7 +42,6 @@ const SchedulePage = () => {
 
     return (
         <div className='flex flex-col h-screen w-screen pl-[240px]'>
-            <Aside />
             <div className={`mt-32 text-3xl ${isHelpOpen && helpStep === 0 ? 'bg-white z-40 rounded-[10px]' : ''}`}>
                 <div className='grid grid-cols-4 mb-4 text-center hidden lg:grid'>
                     <h2>날짜</h2>
