+++ frontend/src/pages/guardians/ViewEmotionAnalysis/ViewEmotionAnalysisPage.js
@@ -1,10 +1,9 @@
 import React, { useContext } from 'react';
-import Aside from '../../../components/section/Aside/Aside';
 import EmotionAnalysisGraph from '../../../components/guardians/EmotionAnalysis/EmotionAnalysisGraph';
 import EmotionAnalysisBoard from '../../../components/guardians/EmotionAnalysis/EmotionAnalysisBoard';
 import Calendar from '../../../components/Calendar/Calendar';
 import Guardian from '../../../components/guardians/Guardian/Guardian';
-import { AppContext } from '../../../App';
+import { AppContext } from '../../../AppContext';
 
 const ViewEmotionAnalysisPage = () => {
     const { isGuardian, setIsGuardian, isWardSetting,
@@ -22,7 +21,6 @@ const ViewEmotionAnalysisPage = () => {
 
     return (
         <div className='flex flex-col h-screen w-screen pl-[240px] pr-10'>
-            <Aside/>
             <div className='pt-28 pl-4 relative h-full'>
                 <div className='flex justify-between text-2xl z-40'>
                     <Calendar/>
